Add Ansible configuration for Buildbot master
Add new Ansible role and playbook.
Move Buildbot configuration from Docker to Ansible.
Add container startup script.
Use Ansible variables and environment variables in configuration.
Use our internal email server instead of an external relay

Add codebase generator function. It takes information about a change,
like repository, branch, and revision, and creates a predictable string
describing the code. This is used to decide upon the builder for a
change and also to tie a fork repository back to its parent for merge
requests.
