+++ ansible-playbook/group_vars/all/buildbot.yml
@@ -1,0 +1,3 @@
+---
+buildbot_version: 1.6.0
+buildbot_worker_port: 9989
