+++ ansible-playbook/group_vars/all/credentials.yml
@@ -3,6 +3,9 @@
 ansibleremote_gitlab_password:       '{{ lookup("password", playbook_dir + "/credentials/ansibleremote/gitlab"                , length=20) }}'
 ansibleremote_root_public_key:       '{{                    playbook_dir + "/credentials/ansibleremote/root_public_key" }}'
 ansibleremote_user_public_key:       '{{                    playbook_dir + "/credentials/ansibleremote/user_public_key" }}'
+buildbot_admin_password:             '{{ lookup("password", playbook_dir + "/credentials/buildbot/admin"                      , length=20) }}'
+buildbot_email_password:             '{{ lookup("password", playbook_dir + "/credentials/buildbot/buildbot@novatech-llc.com"  , length=20) }}'
+buildbot_master_sshkey:              '{{                    playbook_dir + "/credentials/buildbot/master_rsa"                              }}'
 buildsystem_database_build_password: '{{ lookup("password", playbook_dir + "/credentials/buildsystem/database/readuser") }}'
 buildsystem_database_root_password:  '{{ lookup("password", playbook_dir + "/credentials/buildsystem/database/root"           , length=20) }}'
 buildsystem_ssh_prod_user_password:  '{{ lookup("password", playbook_dir + "/credentials/buildsystem/ssh/prod"                , length=20) }}'
