+++ ansible-playbook/roles/buildbot-master/files/master.cfg
@@ -1,0 +1,137 @@
+# -*- python -*-
+# ex: set syntax=python:
+
+# This is a sample buildmaster config file. It must be installed as
+# 'master.cfg' in your buildmaster's base directory.
+
+import os
+from buildbot.plugins import secrets, steps, util
+from buildbot.util import giturlparse
+
+os.environ['LONG_RUN_TIMEOUT'] = '14400'
+
+# This is the dictionary that the buildmaster pays attention to. We also use
+# a shorter alias to save typing.
+c = BuildmasterConfig = {}
+c['workers'] = []
+c['schedulers'] = []
+c['builders'] = []
+c['change_source'] = []
+
+# Defines the TCP port to listen on for connections from workers.
+# This must match the value configured into the buildworkers (with their
+# --master option)
+c['protocols'] = {"pb": {"port": int(os.getenv('BUILDBOT_WORKER_PORT', 9989))}}
+
+# STATUS TARGETS
+
+# 'status' is a list of Status Targets. The results of each build will be
+# pushed to these targets. buildbot/status/*.py has a variety to choose from,
+# including web pages, email senders, and IRC bots.
+
+c['services'] = []
+
+from buildbot.plugins import reporters
+irc = reporters.IRC(host="172.16.64.3",
+                    nick="orionbb",
+                    channels=["#orion"])
+c['services'].append(irc)
+
+mn = reporters.MailNotifier(
+    fromaddr="buildbot@novatechweb.com",
+    sendToInterestedUsers=True,
+    extraRecipients=["andrew.cooper@novatechweb.com"],
+    useTls=True,
+    relayhost=os.getenv('BUILDBOT_SMTP_HOST'),
+    smtpPort=int(os.getenv('BUILDBOT_SMTP_PORT')),
+    smtpUser=os.getenv('BUILDBOT_SMTP_USER'),
+    smtpPassword=os.getenv('BUILDBOT_SMTP_PASS')
+)
+c['services'].append(mn)
+
+# WEB SERVER
+c['www'] = {
+    'port': int(os.getenv('BUILDBOT_WEB_PORT', 8010)),
+    'plugins': {
+        'waterfall_view': True,
+        'console_view': True,
+        'grid_view': True,
+    },
+    'auth': util.UserPasswordAuth([("georgem", "tgbyhn")]),
+    'authz': util.Authz(
+        allowRules=[
+            util.AnyControlEndpointMatcher(role="admins"),
+        ],
+        roleMatchers=[
+            util.RolesFromUsername(roles=['admins'], usernames=["georgem"]),
+        ],
+    ),
+}
+
+# PROJECT IDENTITY
+
+# the 'title' string will appear at the top of this buildbot
+# installation's html.WebStatus home page (linked to the
+# 'titleURL') and is embedded in the title of the waterfall HTML page.
+
+c['title'] = "NovaTech CI"
+c['titleURL'] = "http://git.novatech-llc.com/"
+
+# the 'buildbotURL' string should point to the location where the buildbot's
+# internal web server (usually the html.WebStatus page) is visible. This
+# typically uses the port number set in the Waterfall 'status' entry, but
+# with an externally-visible host name which the buildbot cannot figure out
+# without some help.
+
+c['buildbotURL'] = os.getenv('BUILDBOT_WEB_URL', "http://localhost:%s/" % (c['www']['port']))
+
+# DB URL
+
+c['db'] = {
+    # This specifies what database buildbot uses to store its state.  You can leave
+    # this at its default for all but the largest installations.
+    'db_url' : os.getenv('BUILDBOT_DATABASE', "sqlite://"),
+}
+
+# Extra Options
+c['buildbotNetUsageData'] = 'basic'
+
+def codebaseGenerator(chdict):
+    """ Generate a codebase description.
+
+    The codebase generator takes information about a change, like repository,
+    branch, and revision, and creates a predictable string describing the code.
+    This is used to decide upon the builder for a change and also to tie a fork
+    repository back to its parent for merge requests.
+    """
+
+    giturl = giturlparse(chdict['repository'])
+    return '/'.join((giturl.owner, giturl.repo))
+
+    # path = urlparse(chdict['revlink']).path.lstrip('/')
+
+    # components = []
+    # while path != "":
+    #     path, tail = os.path.split(path)
+    #     if tail != "":
+    #         components.append(tail)
+    # components.reverse()
+    # codebase = os.path.join(*components[0:2])
+    # return codebase
+
+
+c['codebaseGenerator'] = codebaseGenerator
+
+import openembedded
+reload(openembedded)
+c['workers'].extend(openembedded.BuildmasterConfig['workers'])
+c['schedulers'].extend(openembedded.BuildmasterConfig['schedulers'])
+c['builders'].extend(openembedded.BuildmasterConfig['builders'])
+c['change_source'].extend(openembedded.BuildmasterConfig['change_source'])
+
+import ptxdist
+reload(ptxdist)
+c['workers'].extend(ptxdist.BuildmasterConfig['workers'])
+c['schedulers'].extend(ptxdist.BuildmasterConfig['schedulers'])
+c['builders'].extend(ptxdist.BuildmasterConfig['builders'])
+c['change_source'].extend(ptxdist.BuildmasterConfig['change_source'])
