+++ ansible-playbook/roles/buildbot-master/handlers/main.yml
@@ -1,0 +1,2 @@
+---
+# handlers file for buildbot-master
