+++ ansible-playbook/roles/buildbot-master/tests/inventory
@@ -1,0 +1,2 @@
+localhost
+
