+++ ansible-playbook/roles/buildbot-master/vars/main.yml
@@ -1,0 +1,2 @@
+---
+# vars file for buildbot-master
