+++ docker/buildbot-master/buildbot/.gitconfig
@@ -1,0 +1,3 @@
+[user]
+	name = BuildBot
+	email = buildbot@novatechweb.com
