+++ docker/buildbot-master/buildbot/start.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Run original buildbot entrypoint
+exec "/usr/src/buildbot/docker/start_buildbot.sh"
